Teach JWT token to store the user's role
This allows UserController#getCurrentUser(...) to return the role that
the user is currently signed into.

There's another way of implementing this feature, that is to store a new
field currentRole in User. However:

1. If we make it non-persistent (not stored in database), after every
API call, the currentRole will be reset to null. Therefore this doesn't
work.

2. If we make it persistent, the database will store the currentRole
even though the user has logged out (logging out currently doesn't sync
with back-end). This means that the database will be storing additional
redundant details, which isn't as good as the proposed solution since
when the user logs out / ends the session, the JWT will be flushed from
the browser and the currentRole will be flushed away as well.