UserPrincipal: Shift verification of input to setter
UserPrincipal's constructor throws a RuntimeException
RoleNotFoundException if `role` is not in the user's list of roles. This
verification of role is performed in a separate location from the
verification of password.

This is bad software design.

Let's redo the logic to allow the verification of both roles and
passwords to be at NricPasswordRoleAuthenticationProvider.